+++ src/main/java/com/snowflake/kafka/connector/internal/streaming/schemaevolution/SchemaEvolutionTargetItems.java
@@ -1,4 +1,4 @@
-package com.snowflake.kafka.connector.internal.streaming.schematization;
+package com.snowflake.kafka.connector.internal.streaming.schemaevolution;
 
 import java.util.Collections;
 import java.util.List;
