+++ src/main/java/com/snowflake/kafka/connector/internal/streaming/schemaevolution/TableSchema.java
@@ -1,4 +1,4 @@
-package com.snowflake.kafka.connector.internal.streaming.schematization;
+package com.snowflake.kafka.connector.internal.streaming.schemaevolution;
 
 import java.util.Map;
 import java.util.Objects;
