+++ src/main/java/com/snowflake/kafka/connector/internal/streaming/schemaevolution/iceberg/IcebergColumnTypeMapper.java
@@ -1,5 +1,6 @@
-package com.snowflake.kafka.connector.internal.streaming.schematization;
+package com.snowflake.kafka.connector.internal.streaming.schemaevolution.iceberg;
 
+import com.snowflake.kafka.connector.internal.streaming.schemaevolution.ColumnTypeMapper;
 import org.apache.kafka.connect.data.Schema;
 
 public class IcebergColumnTypeMapper extends ColumnTypeMapper {
